Move sliceIntoSubRequests into reindex
