+++ modules/reindex/src/test/java/org/elasticsearch/index/reindex/ReindexParallelizationHelperTests.java
@@ -1,0 +1,63 @@
+/*
+ * Licensed to Elasticsearch under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.elasticsearch.index.reindex;
+
+import org.elasticsearch.action.search.SearchRequest;
+import org.elasticsearch.index.mapper.UidFieldMapper;
+import org.elasticsearch.search.builder.SearchSourceBuilder;
+import org.elasticsearch.test.ESTestCase;
+
+import java.io.IOException;
+
+import static java.util.Collections.emptyList;
+import static org.elasticsearch.index.reindex.ReindexParallelizationHelper.sliceIntoSubRequests;
+import static org.elasticsearch.search.RandomSearchRequestGenerator.createSearchRequest;
+import static org.elasticsearch.search.RandomSearchRequestGenerator.createSearchSourceBuilder;
+
+public class ReindexParallelizationHelperTests extends ESTestCase {
+    public void testSliceIntoSubRequests() throws IOException {
+        SearchRequest searchRequest = createSearchRequest(() -> createSearchSourceBuilder(
+                () -> null,
+                () -> null,
+                () -> null,
+                () -> emptyList()));
+        if (searchRequest.source() != null) {
+            // Clear the slice builder if there is one set. We can't call sliceIntoSubRequests if it is.
+            searchRequest.source().slice(null);
+        }
+        int times = between(2, 100);
+        String field = randomBoolean() ? UidFieldMapper.NAME : randomAsciiOfLength(5);
+        int currentSliceId = 0;
+        for (SearchRequest slice : sliceIntoSubRequests(searchRequest, field, times)) {
+            assertEquals(field, slice.source().slice().getField());
+            assertEquals(currentSliceId, slice.source().slice().getId());
+            assertEquals(times, slice.source().slice().getMax());
+
+            // If you clear the slice then the slice should be the same request as the parent request
+            slice.source().slice(null);
+            if (searchRequest.source() == null) {
+                // Except that adding the slice might have added an empty builder
+                searchRequest.source(new SearchSourceBuilder());
+            }
+            assertEquals(searchRequest, slice);
+            currentSliceId++;
+        }
+    }
+}
